new notebook + new docs

+++ src/polymesh/topo/topo.py
@@ -383,8 +383,8 @@ def cells_at_nodes(topo: TopoLike, *args, frmt=None, assume_regular=False,
             return counts, csr
         return csr
     elif frmt == 'dicts':
-        cellIDs = np.arange(len(topo)) if cellIDs is None else cellIDs
-        nodeIDs = np.arange(len(counts)) if nodeIDs is None else nodeIDs
+        cellIDs = np.arange(len(topo)).astype(np.int32) if cellIDs is None else cellIDs
+        nodeIDs = np.arange(len(counts)).astype(np.int32) if nodeIDs is None else nodeIDs
         ereg, nreg = _nodal_cell_data_to_dicts_(
             counts, ereg, nreg, cellIDs, nodeIDs)
         if return_counts:
